Se agregan clases negocio y visualizacion de listados en la capa TurnosAppWeb

+++ TurnosDominio/EstadoTurno.cs
@@ -8,7 +8,7 @@ namespace TurnosDominio
 {
     public class EstadoTurno
     {
-        public int id { get; set; }
+        public Int16 id { get; set; }
         public string descripcion { get; set; }
         public bool activo { get; set; }
         public override string ToString()
